Get tests and examples working for GO = long long
This commit changes many of the test files under Adapters/ClassicDTK
to get the tests to compile and run using GO = long long. Most of the
changes are very strightforward, also some edits had to be made to
work with the current STK.

All but two tests pass on my Mac using Apple LLVM version 10.0.0 (clang-1000.11.45.5),
on my OSX 10.13.6 laptop:

1:check_no_trailing_whitespaces_or_tabs
34:DataTransferKitOperators_L2Projection_test_MPI_4

There were no whitespace issues, but test #1 uses (BSD) grep on the Mac and that
exits with an exit code of 2 even though it does not locate any whitespace or exits with an error.
I suspect that the nonzero exit code is due to the warning message that it printed:

grep: data: Is a directory

The grep command used is supposed to ignore the data directory, but it does not.

The second test failure is more insidious:

 Tpetra::CrsGraph<int, long long, Kokkos::Compat::KokkosDeviceWrapperNode<Kokkos::Serial, Kokkos::HostSpace> >::insertGlobalIndicesImpl: Proc 0: Not enough capacity to insert 8 indices into local row 11, which currently has 0 entries and total allocation size 0.  Input global column indices: [148, 149, 159, 158, 228, 229, 239, 238], Current global column indices: []
 NOTE: Unit test failed on all processes!